Add converter.py to accelerate-tensorflow-lite
I added this file to the package.yaml, and to my dismay, it complaned
about it not being in the accelerate-tensorflow-lite folder. I think
that (for now), this means that running a program that uses the library
requires to be run from the accelerate-tensorflow-lite directory to set
the cwd correctly. I'm not sure how to verify that stack knows where
this file is, or how to verify where stack puts it.