Make a quick fix for tensor shapes
There's a lot of issues with the current system, however. The input
tensors work quite well as it stands; however, we do need to supply
their shapes separately, and in the correct order. For output tensors,
the situation is worse, as in the current solution, output tensors need
to have the same shape as the first input tensor.

To fix this we will likely have to do some shape inference.