Complete base test suite on accelerate-tensorflow
There are not that many tests yet, but the main bulk of the work is
done. Turns out that it took me forever to figure out how to do it
because the map function wasn't working and I first started trying to do
this with map (+1). Oh, well...

The build files have changed not only to accomodate for the tests, but
there is a little bit of change here and there because the build wasn't
working for me. This should alter the order of linkage which shouldn't
break anything that was already working.