Fix transpose requiring backpermute
The original code that checked whether a backpermute was a transpose did
not work correctly with respect to the transpose function. This commit
should fix the transpose detection in a call to backpermute, but it
might also flag too many things as a transpose; there is also a TODO
comment that contains what might need to change in such an event.